Fix the data issue, in theory have replication file.
What was the issue? Something with the way that I was joining the new installation data with the old.
I'm still not sure what was going on, but what I'm going to do for now is not include the data from
the first quarter of 2025. A future version of this work may (should) include it.